// apps/frontend/src/shared/api/schema.d.ts
export interface components {
    schemas: {
// ...
        UserIdentityDto: {
            /** Format: uuid */
            id: string;
            /** @enum {string} */
            role: "ADMIN" | "STUDENT";
            /** Format: email */
            email?: string;
            fullName?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}